last update used to generate textures and uv coords

+++ LUTGen/Program.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Xml.Linq;
@@ -16,8 +17,31 @@ namespace LUTGen
             }
 
             XDocument doc = XDocument.Load(File.OpenRead("fontinfo.xml"));
-            var charmap = FontInfoReader.Read(doc).ToList();
-            File.WriteAllText($"unicode_font_lut.h", new LUTGenerator(charmap).ToString());
+            if (!File.Exists("charset.csv"))
+            {
+                var charmap = FontInfoReader.Read(doc).ToList();
+                File.WriteAllText($"unicode_font_lut.h", new LUTGenerator(charmap).ToString());
+            }
+            else
+            {
+                var charset = new List<char>();
+                var charmap = FontInfoReader.Read(doc).ToList();
+                var x = File.ReadAllText("charset.csv").Split(",");
+                   
+                foreach (string c in x)
+                {
+                    try
+                    {
+                        charset.Add((char)Int32.Parse(c));
+                    } catch
+                    {
+
+                    }
+                }
+
+                File.WriteAllText($"unicode_font_lut.h", new CharsetLutGenerator(charmap, charset).ToString());
+
+            }
         }
     }
 }
